+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 1.0.3
+
+* Added error checks when loading macro file, Fixed some error messages.
+
+## 1.0.2
+
+* Update package.json & fix README
+
 ## 1.0.1
 
 * Fix README